Fix SF bug #2864996.
It seems that AWT doesn't like when you try to call methods that modify the GUI from within a paintComponent() implementation. In this case, trying to get the file's icon might cause a login dialog to be shown to initialize a VFS session; that really confused AWT, not allowing it to make progress. Result: unusable UI.

The fix does two things: cache the VFSFile instance in VPTFile so that we don't have to fetch it all the time. And allow lazy initialization of the file instance, so that the node's paint code avoids the situation above. There's a visible side-effect to this, which is the tree will "flicker" a little before all file instances are initialized. This is caused by the renderer seeing a null file, which means "file does not exist", before initialization happens.